From: Andrew Zhezherun <zhezherun@gmail.com>   To: leonb@users.sourceforge.net
1094   while (x < dmax)
1095   {
1096     d[x] = s[smax];
1097     x++;
1098   }

I have looked at this code, and from the way it is used it seems that the
interval from smin to smax is open-ended on the rigth side: [smin, smax), so
that smax denotes a position right after the end of the block. In this case
shouldn't this line read rather like the following?

1096     d[x] = s[smax - 1];

When I change the line to this, the problem disappears with no apparent
visual changes. Neither of DjView or Lizardtech plugin have problems on this
file, but I guess they don't use the same compiler as WinDjView. Could you
let me know what you think?