Fix dieze in annotations (#3507031)
See also debian bug #664280.
The new miniexp reserves the character dieze but colors in annotations
are represented by symbols such as #FF00FF. The fix is to introduce
an array of parsing function p_diezechar that are called when following
a dieze character. When this array is zero or when it contains zeroes
(the default), the dieze character is no longer treated specially.